:recycle: move sas_for_impersonation into scanner.py
